sesiones,editar perfil,actualizar clave,recuperar clave,manejo de errores

+++ src/main/java/com/tienda/licorera/sevicio/CategoriaServImpl.java
@@ -2,15 +2,19 @@ package com.tienda.licorera.sevicio;
 
 import java.util.List;
 
+import com.tienda.licorera.error.ErrorServicio;
 import com.tienda.licorera.modelo.Categoria;
+import com.tienda.licorera.modelo.Licor;
 import com.tienda.licorera.repositorio.RepositorioCategoria;
+import com.tienda.licorera.repositorio.RepositorioLicor;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 /**
  * 
- * Esta clase se encarga de implementar los servicios de la interface ICategoriaServicio
+ * Esta clase se encarga de implementar los servicios de la interface
+ * ICategoriaServicio
  * 
  * @author Beelz
  */
@@ -20,59 +24,63 @@ public class CategoriaServImpl implements ICategoriaServicio {
     @Autowired
     private RepositorioCategoria categoriaRepositorio;
 
-    /**
-     * Metodo que busca una categoria en el repositorio
-     * 
-     * @param cod_cat Recibe como parametro el codigo de la categoria a buscar
-     * @return retorna el objeto categoria encontrado segun su codigo
-     * @author Beelz
-     */
+    
+    @Autowired
+    private RepositorioLicor licorRepositorio;
+
     @Override
     public Categoria buscarCat(int cod_cat) {
-
         return categoriaRepositorio.findById(cod_cat).orElse(null);
     }
 
-    /**
-     * Metodo que elimina una categoria del repositorio
-     * 
-     * @param cod_cat Recibe como parametro el codigo de la categoria a eliminar
-     * @author Beelz
-     */
     @Override
-    public void eliminarCat(int cod_cat) {
-        categoriaRepositorio.deleteById(cod_cat);
+    public void eliminarCat(int cod_cat) throws ErrorServicio{
+        List<Licor> listadoLicores = (List<Licor>)licorRepositorio.findByCod_cat(cod_cat);
+        System.out.print(listadoLicores.toString());
+        if (!listadoLicores.isEmpty()) {
+            throw new ErrorServicio("Error al eliminar, Esta categoria contiene licores.");
+        } else if (listadoLicores.isEmpty()) {
+            categoriaRepositorio.deleteById(cod_cat);
+        }
     }
-    /**
-     * Metodo que guarda una categoria en el repositorio
-     * 
-     * @param categoria Recibe como parametro el objeto categoria a guardar
-     * @author Beelz
-     */
+
     @Override
-    public void guardarCategoria(Categoria categoria) {
-        categoriaRepositorio.save(categoria);
+    public Categoria actualizarCategoria(Categoria categoria) throws ErrorServicio{
+        categoria.setNomb_cat(categoria.getNomb_cat().toUpperCase());
+        validarCategoria(categoria);
+        return categoriaRepositorio.save(categoria);
     }
-    /**
-     * Metodo que lista las categorias existentes en el repositorio
-     * 
-     * @return retorna la lista de categorias 
-     * @author Beelz
-     */
+
+    @Override
+    public Categoria guardarCategoria(Categoria categoria) throws ErrorServicio {
+        categoria.setNomb_cat(categoria.getNomb_cat().toUpperCase());
+        validarCategoria(categoria);
+        return categoriaRepositorio.save(categoria);
+    }
+
     @Override
     public List<Categoria> listarTodas() {
 
         return (List<Categoria>) categoriaRepositorio.findAll();
     }
-    /**
-     * Metodo que busca por nombre una categoria existente en el repositorio
-     * 
-     * @param nomb_cat Recibe como parametro el nombre de la categoria a buscar
-     * @return retorna el nombre de la categoria (si la encuentra) 
-     * @author Beelz
-     */
+
     @Override
     public Categoria buscarPorNombre(String nomb_cat) {
         return categoriaRepositorio.findByNomb_cat(nomb_cat);
+    }
+
+    @Override
+    public void validarCategoria(Categoria categoria) throws ErrorServicio {
+
+        // validacion del nombre de la categoria
+        Categoria cat = buscarPorNombre(categoria.getNomb_cat());
+        if (categoria.getNomb_cat() == null || categoria.getNomb_cat().isEmpty()) {
+            throw new ErrorServicio("El campo categoria no puede estar vacio, ingrese la categoria.");
+        } else if (categoria.getNomb_cat().length() > 45) {
+            throw new ErrorServicio("El campo categoria debe contener maximo 45 caracteres.");
+        } else if (cat != null && categoria.getCod_cat() != cat.getCod_cat()) {
+            throw new ErrorServicio(
+                    "El nombre de la categoria: " + categoria.getNomb_cat() + " ya se encuentra registrado");
+        }
     }
 }
